# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: 库存调拨规划器 API
  version: 1.0.0
  description: |
    库存调拨规划器提供完整的调拨申请、库存约束检查、调拨方案生成与确认、多仓自动分配、
    仓库管理、SKU 管理、库存管理及调拨记录查询等能力。

    **主要模块：**
    - 仓库管理 (`/api/warehouses`)
    - SKU 管理 (`/api/skus`)
    - 库存管理 (`/api/inventory`)
    - 调拨申请 (`/api/transfer-requests`)
    - 调拨方案 (`/api/transfer-plans`)
    - 调拨记录 (`/api/transfer-records`)

    **约束分级：**
    - `FATAL`：不可通过，阻断调拨
    - `WARNING`：可确认但需业务复核
    - `INFO`：提示性信息

    **调拨方案状态机：**
    - `PENDING`（待确认）→ `CONFIRMED`（已确认）→ `COMPLETED`（已完成）
    - `PENDING` → `REJECTED`（已驳回）
    - 任意状态 → `CANCELLED`（已取消）
  contact:
    name: Inventory Ops Team
servers:
  - url: http://localhost:3000
    description: 本地开发环境
  - url: /
    description: 当前部署
...
paths:
...
  /api/transfer-records:
    get:
      tags: [Transfer Records]
      summary: 查询调拨记录（分页、多维度筛选）
      operationId: listTransferRecords
      parameters:
        - name: source_warehouse_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: target_warehouse_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: sku_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: plan_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: request_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: start_date
          in: query
          schema: { type: string, format: date }
          description: 操作开始日期（含）
        - name: end_date
          in: query
          schema: { type: string, format: date }
          description: 操作结束日期（含）
        - name: keyword
          in: query
          schema: { type: string, maxLength: 50 }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize50'
      responses:
        '200': { description: 分页记录 }
  /api/transfer-records/summary:
    get:
      tags: [Transfer Records]
      summary: 调拨汇总统计（按方向分组）
      operationId: getTransferRecordSummary
      parameters:
        - name: warehouse_id
          in: query
          schema: { type: integer, minimum: 1 }
        - name: start_date
          in: query
          schema: { type: string, format: date }
        - name: end_date
          in: query
          schema: { type: string, format: date }
        - name: group_by
          in: query
          schema: { type: string, enum: [warehouse, sku], default: warehouse }
          description: 聚合维度
      responses:
        '200':
          description: 入/出方向汇总数据
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      outbound:
                        type: array
                        items: { type: object }
                      inbound:
                        type: array
                        items: { type: object }
  /api/transfer-records/{id}:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    get:
      tags: [Transfer Records]
      summary: 查询单条调拨记录详情
      operationId: getTransferRecord
      responses:
        '200': { description: 记录详情（含仓库/SKU/方案/申请关联信息） }
        '404': { $ref: '#/components/responses/NotFound' }
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema: { type: integer, minimum: 1, example: 1 }
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
...
    PageSize50:
      name: pageSize
      in: query
      schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
...
  responses:
...
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorBody'
...
  schemas:
...
    ErrorBody:
      type: object
      required: [error, code]
      properties:
        error: { type: string, description: 面向用户的错误消息 }
        code:
          type: string
          description: 稳定的机器可识别错误码
          example: WAREHOUSE_NOT_FOUND
        requestId:
          type: string
          description: 请求追踪 ID，用于日志关联
        timestamp: { type: string, format: date-time }
...